use `async fn` for async trait
for readability

A Send variant may be nice to use with Embassy's SendSpawner.
However, embedded-hal-async traits don't have Send variants, so
it's unlikely the that the drivers called in the impl's of
this trait would have async fn's that return Send Futures.
You probably don't want to be writing different frames to one LED
device from multiple threads simultaneously anyway; that would
require the driver to implement synchronization to avoid
interleaving of separate frames, which would add a run time cost.