add ExactSizeIterator trait bound
This allows implementations to automatically handle protocol
details that depend on the number of LEDs addressed. This is
required for sending the correct length for the end frame in the
APA102/SK9822 protocol:
https://github.com/smart-leds-rs/apa102-spi-rs/pull/18

AFAIK the only downside to this is that core::iter::Chain doesn't
implement ExactSizeIterator. I don't think that generally matters
for downstream users, who are most likely passing slices of LED
data to SmartLedsWrite::write.